+++ lizard_graph/views.py
@@ -171,7 +171,7 @@ def cached_time_series_aggregated(graph_item, start, end,
         graph_item, start, end, aggregation, aggregation_period):
 
         return ('ts_agg::%s:%s:%s:%s:%s:%s::%s:%s:%s:%s' % (
-            graph_item.fews_norm_db_name, graph_item.location,
+            graph_item.fews_norm_source.database_name, graph_item.location,
             graph_item.parameter, graph_item.module, graph_item.time_step,
             graph_item.qualifierset, start, end, aggregation,
             aggregation_period)).replace(' ', '_')
@@ -191,7 +191,7 @@ def cached_time_series_from_graph_item(graph_item, start, end):
     """
     def time_series_key(graph_item, start, end):
         return ('ts::%s:%s:%s:%s:%s:%s::%s:%s' % (
-            graph_item.fews_norm_db_name, graph_item.location,
+            graph_item.fews_norm_source.database_name, graph_item.location,
             graph_item.parameter, graph_item.module, graph_item.time_step,
             graph_item.qualifierset, start, end)).replace(' ', '_')
     cache_key = time_series_key(graph_item, start, end)